# internal/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Gin 的快速启动项目",
        "title": "Gin Admin",
        "contact": {},
        "version": "v1.0.0"
    },
    "paths": {
...
        "/api/v1/auth/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "AuthAPI"
                ],
                "summary": "Get current user info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-po_User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "AuthAPI"
                ],
                "summary": "Update current user info",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthUpdateUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "UserAPI"
                ],
                "summary": "Query user list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pagination size, less than 0 is considered as unlimited quantity. default(20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name of user",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pagination index. default(1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "activated",
                            "freezed"
                        ],
                        "type": "string",
                        "description": "Status of user (activated, freezed)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "total number of items",
                        "name": "total",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Username for login",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Whether to include role IDs",
                        "name": "withRoles",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ResultList-po_User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "UserAPI"
                ],
                "summary": "Create user record",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-po_User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "UserAPI"
                ],
                "summary": "Get user record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-po_User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "UserAPI"
                ],
                "summary": "Update user record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "UserAPI"
                ],
                "summary": "Delete user record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/reset-pwd": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "UserAPI"
                ],
                "summary": "Reset user password by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.AuthUpdateUserReq": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email of user",
                    "type": "string",
                    "maxLength": 128
                },
                "nickName": {
                    "description": "Name of user",
                    "type": "string",
                    "maxLength": 64
                },
                "phone": {
                    "description": "Phone number of user",
                    "type": "string",
                    "maxLength": 32
                },
                "wechat": {
                    "description": "Wechat account",
                    "type": "string",
                    "maxLength": 64
                }
            }
        },
...
        "dto.List-po_User": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/po.User"
                    }
                },
                "limit": {
                    "description": "pagination size, less than 0 is considered as unlimited quantity. default(20)",
                    "type": "integer"
                },
                "page": {
                    "description": "pagination index. default(1)",
                    "type": "integer"
                },
                "total": {
                    "description": "total number of items",
                    "type": "integer"
                }
            }
        },
...
        "dto.Result-any": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.Result-po_User": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/po.User"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.ResultList-po_User": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/dto.List-po_User"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.UserCreateReq": {
            "type": "object",
            "required": [
                "nickName",
                "roleIds",
                "status",
                "username"
            ],
            "properties": {
                "description": {
                    "description": "Description of user",
                    "type": "string",
                    "maxLength": 1024
                },
                "email": {
                    "description": "Email of user",
                    "type": "string",
                    "maxLength": 128
                },
                "nickName": {
                    "description": "Name of user",
                    "type": "string",
                    "maxLength": 64
                },
                "password": {
                    "description": "Password for login (md5 hash)",
                    "type": "string",
                    "maxLength": 64
                },
                "phone": {
                    "description": "Phone number of user",
                    "type": "string",
                    "maxLength": 32
                },
                "realName": {
                    "description": "Real name of user",
                    "type": "string",
                    "maxLength": 64
                },
                "roleIds": {
                    "description": "Roles of user",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "description": "Status of user (activated, freezed)",
                    "type": "string",
                    "enum": [
                        "activated",
                        "freezed"
                    ]
                },
                "username": {
                    "description": "Username for login",
                    "type": "string",
                    "maxLength": 64
                },
                "wechat": {
                    "description": "Wechat account",
                    "type": "string",
                    "maxLength": 64
                }
            }
        },
        "dto.UserUpdateReq": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description of user",
                    "type": "string",
                    "maxLength": 1024
                },
                "email": {
                    "description": "Email of user",
                    "type": "string",
                    "maxLength": 128
                },
                "nickName": {
                    "description": "Name of user",
                    "type": "string",
                    "maxLength": 64
                },
                "password": {
                    "description": "Password for login (md5 hash)",
                    "type": "string",
                    "maxLength": 64
                },
                "phone": {
                    "description": "Phone number of user",
                    "type": "string",
                    "maxLength": 32
                },
                "realName": {
                    "description": "Real name of user",
                    "type": "string",
                    "maxLength": 64
                },
                "roleIds": {
                    "description": "Roles of user",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "description": "Status of user (activated, freezed)",
                    "type": "string",
                    "enum": [
                        "activated",
                        "freezed"
                    ]
                },
                "username": {
                    "description": "Username for login",
                    "type": "string",
                    "maxLength": 64
                },
                "wechat": {
                    "description": "Wechat account",
                    "type": "string",
                    "maxLength": 64
                }
            }
        },
...
        "po.Menu": {
            "type": "object",
            "properties": {
                "children": {
                    "description": "Child menus",
                    "allOf": [
                        {
                            "$ref": "#/definitions/po.Menus"
                        }
                    ]
                },
                "component": {
                    "description": "Component path of view",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Create time",
                    "type": "string"
                },
                "extra": {
                    "description": "Extra data for frontend",
                    "type": "object",
                    "additionalProperties": {}
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "method": {
                    "description": "Http method of resource",
                    "type": "string"
                },
                "name": {
                    "description": "Display name of menu",
                    "type": "string"
                },
                "parentId": {
                    "description": "Parent ID (From Menu.ID)",
                    "type": "string"
                },
                "path": {
                    "description": "Access path of menu",
                    "type": "string"
                },
                "rank": {
                    "description": "Rank for sorting (Order by desc)",
                    "type": "integer"
                },
                "redirect": {
                    "description": "Redirect path of menu",
                    "type": "string"
                },
                "roles": {
                    "$ref": "#/definitions/po.Roles"
                },
                "status": {
                    "description": "Status of menu (enabled, disabled)",
                    "type": "string"
                },
                "title": {
                    "description": "Menu title",
                    "type": "string"
                },
                "type": {
                    "description": "Type of menu (catalog, menu, button)",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Update time",
                    "type": "string"
                }
            }
        },
        "po.Menus": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/po.Menu"
            }
        },
        "po.Role": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code of role (unique)",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Create time",
                    "type": "string"
                },
                "description": {
                    "description": "Details about role",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/po.Menu"
                    }
                },
                "name": {
                    "description": "Display name of role",
                    "type": "string"
                },
                "rank": {
                    "description": "Rank for sorting",
                    "type": "integer"
                },
                "status": {
                    "description": "Status of role (disabled, enabled)",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Update time",
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/po.User"
                    }
                }
            }
        },
        "po.Roles": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/po.Role"
            }
        },
        "po.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "Avatar URL",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Create time",
                    "type": "string"
                },
                "description": {
                    "description": "Details about user",
                    "type": "string"
                },
                "email": {
                    "description": "Email of user",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "nickName": {
                    "description": "Name of user",
                    "type": "string"
                },
                "phone": {
                    "description": "Phone number of user",
                    "type": "string"
                },
                "realName": {
                    "description": "Real name of user",
                    "type": "string"
                },
                "roles": {
                    "description": "Roles of user",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/po.Role"
                    }
                },
                "status": {
                    "description": "Status of user (activated, freezed)",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Update time",
                    "type": "string"
                },
                "username": {
                    "description": "Username for login",
                    "type": "string"
                },
                "wechat": {
                    "description": "Wechat account",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}